Refactor/rename component helpers
Aligning with new 2.0 structure where anything not used externally should be within the `helpers` directory for each resource component.
