+++ src/components/content/allergies/helpers/column.tsx
