+++ src/components/content/document/helpers/filters.tsx
