+++ src/fhir/document.ts
@@ -1,6 +1,6 @@
 import { searchCommonRecords } from "./search-helpers";
 import { useQueryWithPatient } from "..";
-import { applyDocumentFilters } from "@/components/content/document/patient-document-filters";
+import { applyDocumentFilters } from "@/components/content/document/helpers/filters";
 import { orderBy } from "@/utils/nodash";
 import { QUERY_KEY_PATIENT_DOCUMENTS } from "@/utils/query-keys";
 import { Telemetry, withTimerMetric } from "@/utils/telemetry";
